+++ src/icalendar/tests/test_unit_caselessdict.py
@@ -59,7 +59,7 @@ class TestCaselessdict(unittest.TestCase):
 
     def test_caselessdict_copy(self):
         CaselessDict = icalendar.caselessdict.CaselessDict
-        
+
         original_dict = CaselessDict(key1='val1', key2='val2')
         copied_dict = original_dict.copy()