+++ src/icalendar/tests/__init__.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 # unittest/unittest2 importer
 import unittest
 if not hasattr(unittest.TestCase, 'assertIsNotNone'):