+++ src/icalendar/tests/test_multiple.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from icalendar import Calendar
 from icalendar.prop import vText
 from icalendar.tests import unittest