+++ src/icalendar/tests/test_unit_cal.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from datetime import datetime
 from datetime import timedelta
 from icalendar.tests import unittest