+++ CHANGES.rst
@@ -1,11 +1,16 @@
-
 Changelog
 =========
 
 3.9.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+New:
+
+- *add item here*
+
+Fixes:
+
+- *add item here*
 
 
 3.9.1 (2015-09-08)