+++ CHANGES.rst
@@ -10,6 +10,10 @@ New:
 
 Fixes:
 
+- Fixed cal.Component.from_ical() representing an unknown component
+  as one of the known.
+  [stlaz]
+
 - Fixed date-time being recognized as date or time during parsing. Added
   better error handling to parsing from ical strings.
   [stlaz]