+++ CHANGES.rst
@@ -13,11 +13,15 @@ Fixes:
 - Fixed possible IndexError exception during parsing of an ical string.
   [stlaz]
 
-- Fixed date-time being recognized as date or time during parsing. Added
-  better error handling to parsing from ical strings.
+- When doing a boolean test on ``icalendar.cal.Component``, always return ``True``.
+  Before it was returning ``False`` due to CaselessDict, if it didn't contain any items.
   [stlaz]
 
-- Added __version__ attribute to init.py
+- Fixed date-time being recognized as date or time during parsing.
+  Added better error handling to parsing from ical strings.
+  [stlaz]
+
+- Added __version__ attribute to init.py.
   [TomTry]
 
 - Documentation fixes.